% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cori.data.verse-package.R
\docType{package}
\name{cori.data.verse-package}
\alias{cori.data.verse}
\alias{cori.data.verse-package}
\title{cori.data.verse: Rural Dataverse by the Center on Rural Innovation (CORI)}
\description{
A universe of rural innovation data, tools, research, and analysis from the Center on Rural Innovation (CORI).
}
